YeildCalculator 테스트에서 yield 타입 double로 변환

+++ src/test/java/lotto/model/YieldCalculatorTest.java
@@ -15,7 +15,7 @@ class YieldCalculatorTest {
         myWallet.saveMoney(10000);
         myWallet.saveWinnings(5000);
 
-        float yield = YieldCalculator.calculateYield(myWallet);
+        double yield = YieldCalculator.calculateYield(myWallet);
 
         assertThat(yield).isEqualTo(50.0f);
     }
@@ -27,7 +27,7 @@ class YieldCalculatorTest {
         myWallet.saveMoney(10000);
         myWallet.saveWinnings(0);
 
-        float yield = YieldCalculator.calculateYield(myWallet);
+        double yield = YieldCalculator.calculateYield(myWallet);
 
         assertThat(yield).isEqualTo(0.0f);
     }
